+++ tests/ttnn/unit_tests/gtests/tensor/test_host_buffer.cpp
@@ -9,8 +9,8 @@
 #include <functional>
 #include <vector>
 
-#include "ttnn/tensor/host_buffer/host_buffer.hpp"
-#include "ttnn/tensor/host_buffer/memory_pin.hpp"
+#include <tt-metalium/host_buffer.hpp>
+#include <tt-metalium/memory_pin.hpp>
 
 namespace tt::tt_metal {
 namespace {
