+++ tests/ttnn/unit_tests/gtests/tensor/test_memory_pin.cpp
@@ -7,7 +7,7 @@
 #include <functional>
 #include <utility>
 
-#include "ttnn/tensor/host_buffer/memory_pin.hpp"
+#include <tt-metalium/memory_pin.hpp>
 
 namespace tt::tt_metal {
 namespace {
