+++ ttnn/cpp/ttnn-pybind/pytensor.cpp
@@ -10,7 +10,7 @@
 #include <memory>
 
 #include "small_vector_caster.hpp"  // NOLINT - for pybind11 SmallVector binding support.
-#include "ttnn/tensor/host_buffer/host_buffer.hpp"
+#include <tt-metalium/host_buffer.hpp>
 #include "ttnn/tensor/tensor.hpp"
 #include <tt-metalium/graph_tracking.hpp>
 #include <tt_stl/overloaded.hpp>
